Add detail field to ImageMessageContent
The ImageMessageContent interface now includes a new field called "detail" which can be set to 'auto', 'high', or 'low'. This allows specifying the desired level of detail when sending an image URL in a message. 'low' does a 512x512 representation of the image. Default is auto. This was added because of an issue with some images not being processed correctly. Some forums indicated this might help. It didn't!