Cleanup file upload API request
The file upload API request was previously changed to no longer use FormData and instead sends the base64-encoded file chunk directly as the request body. This simply cleans up the code that is no longer used or necessary. The original change was made because there was corruption of the data