Add Pica library and its types
This commit adds the Pica library to the dependencies, which is a high-quality image resizing and scaling library. The types for Pica (@types/pica) were also added to the devDependencies for proper type-checking and intellisense support during the development process.

This was included to try to fix images that were not being processed correctly, but it did not fix that issue unfortunately. Still maybe valuable in some cases, should we decide to implement it.