Fix message end ref component size, minor indentation fixes
The div that handled message end ref (used for scrolling to bottom) was too large and taking up a lot of space. It worked in the previous UI layout because of other factors, but here it created a lot of dead space at the bottom of the page. Reducing the size has not affected the functionality in any way that I can detect.