Add downlevel iteration flag
This is necessary to allow iteration over UInt8 arrays, which will be used for recording user audio.

The tradeoff is less performant iterations more generally. That shouldn't be terrible given that most of the application is calling services rather than handling things itself, but needs to be assessed in terms of effect on other features as well.