Add ChatInputUrl component for URL input functionality
This commit introduces the ChatInputUrl component to the ChatInput form. It allows users to add URLs as part of their chat input. Included related state and event handlers for URL input support.

These are basically handled as files so we are able to reuse as much of that functionality as possible and it instantly allows us multi-file / link / image handling at the same time.