Add target language autonyms and scaffold input options
Introduced `autonym` for each language in the dropdown for localization. Added an option to type the target language manually instead of selecting from a dropdown, though this is not fully implemented for UI reasons. Ultimately it would give the user flexibility in choosing the target language beyond a set of standard options.